Strip doc comments back to a minimum
The removed documentation comments were all copied from the FreeType
API documentation, and tweaked in various ways to align more closely
(but not exactly) with the Go API exported by this module. They were
a pain to copy, paste, and then tweak in a satisfactory manner.

It is safer to simply have a minimal description for each type and
function. Keeping the link to the real documentation makes it
relatively easy to reference the canonical documentation.