Use Pascal case for 'Sfnt' in function names
'SFNT', all uppercase, is the canonical use when refering to SFNT
tables. However the freetype api uses 'Sfnt' in function names. Using
the freetype convention more closely aligns with freetype, and
(subjectively) makes the function names more readable. Although the
downside is that it doesn't adhere to Go naming conventions.